refactor(frontend): split flow-graph-utils into route, edge, layout modules
Closes #352. Pure move of helpers from the monolithic flow-graph-utils
file into flow-graph-routes, flow-graph-edges, and flow-graph-layout;
updates imports and renames the subsystem unit test file.
